requirements.txt
torch>=1.13.0
torchvision>=0.14.0
tqdm>=4.64.0
numpy
Pillow
